Modify: readme + error 해결
