+++ src/pages/AddProduct/ProductInfos/ProductName/ProductName.js
@@ -2,21 +2,26 @@ import React, { useState } from 'react';
 import * as S from './ProductName.style';
 
 const ProductName = () => {
-  const [inputValue, setInputValue] = useState('');
+  const [product, setProduct] = useState('');
 
   const handleInput = e => {
+    // const value = nameRef.current.value;
+
     const userInput = e.target.value;
-    setInputValue(userInput);
+    setProduct(userInput);
+    // nameRef.current = userInput;
+    // setInputValue(userInput);
+    console.log(userInput);
   };
-  console.log(inputValue);
 
   return (
     <S.Wrapper>
       <S.Input
         type="text"
-        name={inputValue}
-        value={inputValue}
+        name="dd"
+        value={product}
         onChange={handleInput}
+        // ref={nameRef}
         placeholder="상품명을 입력해주세요"
       />
     </S.Wrapper>
